+++ Utilities/MetaIO/metaObject.h
@@ -56,7 +56,7 @@ class MetaObject
 
       bool  m_BinaryDataByteOrderMSB;
 
-      float m_CompressedDataSize;
+      unsigned int m_CompressedDataSize;
       bool m_WriteCompressedDataSize; // Used internally to set if the dataSize should be written
       bool m_CompressedData;
